Horizontal Scroll Bug
The .card-deck-wrapper class has a negative margin (~20px). Padding on the
container is smaller (~15px) which would cause the wrapper to extend beyond the
page.

Rather than override some stuff, I decided to switch to a fixed width
container instead. No reason for the fluid container on that section
anyway.